+++ src/test/java/org/cicirello/math/rand/EnhancedRandomGeneratorTests.java
@@ -22,10 +22,6 @@
  */
 package org.cicirello.math.rand;
 
-import java.util.random.RandomGenerator;
-import java.util.stream.DoubleStream;
-import java.util.stream.IntStream;
-import java.util.stream.LongStream;
 import java.util.SplittableRandom;
 
 import org.junit.jupiter.api.*;
@@ -387,44 +383,6 @@ public class EnhancedRandomGeneratorTests {
 		assertTrue(result2[1] < result2[2]);
 	}
 	
-	@Test
-	public void testSample() {
-		validateSample((erg, n, k) -> erg.sample(n, k, null));
-	}
-	
-	@Test
-	public void testSamplePool() {
-		validateSample((erg, n, k) -> erg.samplePool(n, k, null));
-	}
-	
-	@Test
-	public void testSampleReservoir() {
-		validateSample((erg, n, k) -> erg.sampleReservoir(n, k, null));
-	}
-	
-	@Test
-	public void testSampleInsertion() {
-		validateSample((erg, n, k) -> erg.sampleInsertion(n, k, null));
-	}
-	
-	@Test
-	public void testSamplePoolProbability() {
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator(new SplittableRandom(42));
-		int n = 100;
-		int[] result = erg.sample(n, 0.05);
-		validateNoRepeats(n, result);
-		assertTrue(result.length < n);
-		
-		result = erg.sample(n, 0.95);
-		validateNoRepeats(n, result);
-		assertTrue(result.length > 0);
-		
-		result = erg.sample(n, 0.5);
-		validateNoRepeats(n, result);
-		assertTrue(result.length > 0);
-		assertTrue(result.length < n);
-	}
-	
 	private int countTrue(boolean[] b) {
 		int count = 0;
 		for (boolean bool : b) {
@@ -450,29 +408,5 @@ public class EnhancedRandomGeneratorTests {
 				assertTrue(Math.abs(result[i]-result[j]) <= window);
 			}
 		}
-	}
-	
-	private void validateNoRepeats(int n, int[] result) {
-		boolean[] in = new boolean[n];
-		for (int e : result) {
-			assertFalse(in[e]);
-			in[e] = true;
-		}
-	}
-	
-	private void validateSample(SampleNK sampler) {
-		EnhancedRandomGenerator erg = new EnhancedRandomGenerator();
-		int n = 6;
-		int[] result = null;
-		for (int k = 1; k <= 6; k++) {
-			result = sampler.apply(erg, n, k);
-			assertEquals(k, result.length);
-			validateNoRepeats(n, result);
-		}
-	}
-	
-	@FunctionalInterface
-	private static interface SampleNK {
-		int[] apply(EnhancedRandomGenerator erg, int n, int k);
 	}
 }
